backend/requirements.txt
flask
flask-cors
chromadb
onnxruntime
anthropic
pypdf
python-docx
python-dotenv
PyMuPDF
Pillow
gunicorn
cryptography
pinecone
